Chaos sanitizer: development
