+++ dbms/src/Common/ThreadFuzzer.h
@@ -1,0 +1,74 @@
+#include <cstdint>
+
+
+namespace DB
+{
+
+/** Allows to randomize thread scheduling and insert various glitches across whole program for testing purposes.
+  * It is done by setting up a timer that will send PROF signal to every thread when certain amount of CPU time has passed.
+  *
+  * To initialize ThreadFuzzer, call ThreadFuzzer::instance().
+  * The behaviour is controlled by environment variables:
+  *
+  * THREAD_FUZZER_CPU_TIME_PERIOD_US  - period of signals in microseconds.
+  * THREAD_FUZZER_YIELD_PROBABILITY   - probability to do 'sched_yield'.
+  * THREAD_FUZZER_MIGRATE_PROBABILITY - probability to set CPU affinity to random CPU core.
+  * THREAD_FUZZER_SLEEP_PROBABILITY   - probability to sleep.
+  * THREAD_FUZZER_SLEEP_TIME_US       - amount of time to sleep in microseconds.
+  *
+  * ThreadFuzzer will do nothing if environment variables are not set accordingly.
+  *
+  * The intention is to reproduce thread synchronization bugs (race conditions and deadlocks) more frequently in tests.
+  * We already have tests with TSan. But TSan only covers "physical" synchronization bugs, but not "logical" ones,
+  *  where all data is protected by synchronization primitives, but we still have race conditions.
+  * Obviously, TSan cannot debug distributed synchronization bugs.
+  *
+  * The motivation for this tool is an evidence, that concurrency bugs are more likely to reproduce
+  *  on bad unstable virtual machines in a dirty environments.
+  *
+  * The idea is not new, see also:
+  * https://channel9.msdn.com/blogs/peli/concurrency-fuzzing-with-cuzz
+  *
+  * Notes:
+  * - it can be also implemented with instrumentation (example: LLVM Xray) instead of signals.
+  * - it's also reasonable to insert glitches around interesting functions (example: mutex lock/unlock, starting of threads, etc.),
+  *   it is doable with wrapping these functions (todo).
+  * - we should also make the sleep time random.
+  * - sleep obviously helps, but the effect of yield and migration is unclear.
+  */
+class ThreadFuzzer
+{
+public:
+    static ThreadFuzzer & instance()
+    {
+        static ThreadFuzzer res;
+        return res;
+    }
+
+    bool isEffective() const
+    {
+        return cpu_time_period_us != 0
+            && (yield_probability > 0
+                || migrate_probability > 0
+                || (sleep_probability > 0 && chaos_sleep_time_us > 0));
+    }
+
+private:
+    uint64_t cpu_time_period_us = 0;
+    double yield_probability = 0;
+    double migrate_probability = 0;
+    double sleep_probability = 0;
+    double chaos_sleep_time_us = 0;
+
+    int num_cpus = 0;
+
+
+    ThreadFuzzer();
+
+    void initConfiguration();
+    void setup();
+
+    static void signalHandler(int);
+};
+
+}
